// src/proto/enums.ts
export enum ResponseType {
  SUCCESS_ATOM = 1,
  SUCCESS_SEQUENCE = 2,
  SUCCESS_PARTIAL = 3,
  WAIT_COMPLETE = 4,
  SERVER_INFO = 5,
  CLIENT_ERROR = 16,
  COMPILE_ERROR = 17,
  RUNTIME_ERROR = 18,
}
export enum ErrorType {
  INTERNAL = 1000000,
  RESOURCE_LIMIT = 2000000,
  QUERY_LOGIC = 3000000,
  NON_EXISTENCE = 3100000,
  OP_FAILED = 4100000,
  OP_INDETERMINATE = 4200000,
  USER = 5000000,
  PERMISSION_ERROR = 6000000,
}